Fix: System-wide integer overflow protection (32-bit long) and optimized visual examples

+++ exported_sketches/acid/acid.ino
@@ -1,4 +1,4 @@
-// MOZZIFLOW v110.9 BALANCED CORE REFINED SKETCH
+// MOZZIFLOW v111.0 BALANCED CORE REFINED SKETCH
 #include <Mozzi.h>
 #include <Oscil.h>
 #include <Metronome.h>
@@ -33,22 +33,22 @@ void updateControl() {
 
 AudioOutput updateAudio() {
     // Control logic moved to audio loop for node aoov
-    if(mozzimetronome_aoov_lastbpm != (float)135) { mozzimetronome_aoov.setBPM((float)135); mozzimetronome_aoov_lastbpm = (float)135;}
+    if(mozzimetronome_aoov_lastbpm != (float)(long)135) { mozzimetronome_aoov.setBPM((float)(long)135); mozzimetronome_aoov_lastbpm = (float)(long)135;}
         node_aoov_out = mozzimetronome_aoov.ready() ? 255 : 0;
     node_5auw_out = mozziead_5auw.next();
     // Control logic moved to audio loop for node 5auw
-    if((int)node_aoov_out>0){ mozziead_5auw.start((unsigned int)10, (unsigned int)200); }
+    if((long)node_aoov_out>0){ mozziead_5auw.start((unsigned int)(long)10, (unsigned int)(long)200); }
     node_3gcd_out = oscil_3gcd.next();
     // Control logic moved to audio loop for node 3gcd
-    oscil_3gcd.setFreq((float)55);
-    mozzimultires_cfnu.next((int)node_3gcd_out);
+    oscil_3gcd.setFreq((float)(long)55);
+    mozzimultires_cfnu.next((int)(long)node_3gcd_out);
         node_cfnu_low = mozzimultires_cfnu.low();
         node_cfnu_high = mozzimultires_cfnu.high();
         node_cfnu_band = mozzimultires_cfnu.band();
         node_cfnu_notch = mozzimultires_cfnu.notch();
     // Control logic moved to audio loop for node cfnu
-    mozzimultires_cfnu.setCutoffFreqAndResonance((uint8_t)node_5auw_out, (uint8_t)220);
-    node_776e_out = ((int)node_cfnu_low >> (int)2);
+    mozzimultires_cfnu.setCutoffFreqAndResonance((uint8_t)(long)node_5auw_out, (uint8_t)(long)220);
+    node_776e_out = ((long)node_cfnu_low >> (int)2);
     return MonoOutput::from8Bit((int)node_776e_out);
 }
