+++ exported_sketches/drums/drums.ino
@@ -1,4 +1,4 @@
-// MOZZIFLOW v110.9 BALANCED CORE REFINED SKETCH
+// MOZZIFLOW v111.0 BALANCED CORE REFINED SKETCH
 #include <Mozzi.h>
 #include <Oscil.h>
 #include <Metronome.h>
@@ -26,15 +26,15 @@ void updateControl() {
 
 AudioOutput updateAudio() {
     // Control logic moved to audio loop for node m1
-    if(mozzimetronome_m1_lastbpm != (float)124) { mozzimetronome_m1.setBPM((float)124); mozzimetronome_m1_lastbpm = (float)124;}
+    if(mozzimetronome_m1_lastbpm != (float)(long)124) { mozzimetronome_m1.setBPM((float)(long)124); mozzimetronome_m1_lastbpm = (float)(long)124;}
         node_m1_out = mozzimetronome_m1.ready() ? 255 : 0;
     node_kick_out = oscil_kick.next();
     // Control logic moved to audio loop for node kick
-    oscil_kick.setFreq((float)55);
+    oscil_kick.setFreq((float)(long)55);
     node_env_out = mozziead_env.next();
     // Control logic moved to audio loop for node env
-    if((int)node_m1_out>0){ mozziead_env.start((unsigned int)20, (unsigned int)200); }
-    node_vca_out = (int)((long)node_kick_out * node_env_out >> 8);
+    if((long)node_m1_out>0){ mozziead_env.start((unsigned int)(long)20, (unsigned int)(long)200); }
+    node_vca_out = ((int)(((long)node_kick_out * (long)node_env_out) >> 8));
     return MonoOutput::from8Bit((int)node_vca_out);
 }
